+++ microservices/notification-service/api/models.py
@@ -8,7 +8,7 @@ class Notification(models.Model):
         ('error', 'Error'),
     ]
 
-    user_id = models.IntegerField()
+    user_id = models.BigIntegerField()
     message = models.TextField()
     notification_type = models.CharField(max_length=20, choices=NOTIFICATION_TYPES, default='info')
     is_read = models.BooleanField(default=False)
